correction bug cookies

+++ BackEnd/api/models/comments.go
@@ -6,8 +6,8 @@ type Comments struct {
 	ID      int64     `json:"id"`
 	Content string    `json:"content"`
 	Date    time.Time `json:"created_at"`
-	Event   int8      `json:"event_id"`
-	User    int8      `json:"user_id"`
+	Event   int64     `json:"event_id"`
+	User    int64     `json:"user_id"`
 }
 
 type CommentsJSON struct {
